- Setup jest testing for the backend
- Unit testing the userService
- StatService should send the userStat item's id
- UserStatDTO should expect the stat to come with a category not categoryId
- Map the stats to the DTO before creating a userStats object with the stat data

+++ backend/services/userService.js
@@ -32,6 +32,11 @@ async function getUserById(id) {
 }
 
 async function deleteUser(id) {
+  const existingUser = await prisma.user.findUnique({
+    where: { id: id },
+  });
+  if (!existingUser) throw new Error(`User with id "${id}" not found.`);
+
   const user = await prisma.user.delete({
     where: { id },
     select: {
@@ -43,13 +48,16 @@ async function deleteUser(id) {
       updatedAt: true,
     },
   });
-
-  if (!user) throw new Error("User not found");
   // return the user that was deleted
   return user;
 }
 
 async function updateUser(id, data) {
+  const existingUser = await prisma.user.findUnique({
+    where: { id: id },
+  });
+  if (!existingUser) throw new Error(`User with id "${id}" not found.`);
+
   const user = await prisma.user.update({
     where: { id },
     data,
@@ -62,7 +70,6 @@ async function updateUser(id, data) {
       updatedAt: true,
     },
   });
-  if (!user) throw new Error("User not found");
   return user;
 }
